split out TagUtils class
